fix: replace assets to public, remove css files, rename routes; feat: add disabling for loging and reg btns after click
