+++ index.html
@@ -4,7 +4,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <link rel="icon" type="image/png" sizes="32x32" href="./src/assets/images/rss-logo.svg">
+  <link rel="icon" type="image/png" sizes="32x32" href="public/images/rss-logo.svg">
   <title>eCommerce Application</title>
 </head>
