--- src/components/modals/confirm-modal/confirm-modal.css
@@ -1,9 +1,0 @@
-.buttons-container {
-  display: flex;
-  justify-content: space-around;
-  margin-top: 20px;
-}
-
-.buttons-container .modal__button {
-  cursor: pointer;
-}
