+++ src/controllers/registration/registration-controller.ts
@@ -2,22 +2,23 @@ import { authService } from '../../commerce-tools/auth-service.ts';
 import { handleApiError } from '../../commerce-tools/models/utils/handle-api-error.ts';
 import { ModalGreeting } from '../../components/modals/modal-greeting.ts';
 import RegistrationModel from '../../model/registration/registration-model.ts';
-import { Layout } from '../../pages/layout.ts';
+import { Layout } from '../../pages/layout/layout.ts';
 import RegistrationPage from '../../pages/registration/index.ts';
-import { route } from '../../router/index.ts';
+import { route } from '../../router';
 import {
   isCommercetoolsApiError,
   isFormName,
   isHTMLCheckboxElement,
   isHTMLInputElement,
   isHTMLSelectElement,
-} from '../../shared/models/typeguards.ts/index.ts';
+} from '../../shared/models/typeguards.ts';
 import { convertInputName } from '../../shared/utils/convert-input-name.ts';
 import { LoginController } from '../login/login-controller.ts';
 
 export class RegistrationController {
   private readonly page: RegistrationPage;
   private readonly model: RegistrationModel;
+  private isActive: boolean = true;
 
   constructor() {
     this.page = RegistrationPage.getInstance();
@@ -41,52 +42,58 @@ export class RegistrationController {
   }
 
   private onClickRegistration = async (): Promise<void> => {
-    const data = {
-      email: this.model.currentFormValues.email,
-      password: this.model.currentFormValues.password,
-      dateOfBirth: this.model.currentFormValues.birthday,
-      firstName: this.model.currentFormValues.name,
-      lastName: this.model.currentFormValues.surname,
-      addresses: [
-        {
-          country: this.model.currentFormValues.country,
-          streetName: this.model.currentFormValues.street,
-          postalCode: this.model.currentFormValues['postalCode'],
-          city: this.model.currentFormValues.city,
-        },
-        {
-          country: this.model.currentFormValues['countryBilling'],
-          streetName: this.model.currentFormValues['streetBilling'],
-          postalCode: this.model.currentFormValues['postalCodeBilling'],
-          city: this.model.currentFormValues['cityBilling'],
-        },
-      ],
-      shippingAddresses: [0],
-      billingAddresses: this.model.currentFormValues['isShippingAsBilling'] ? [0] : [1],
+    if (this.isActive) {
+      this.isActive = false;
 
-      ...(this.model.currentFormValues['isDefaultShipping'] && {
-        defaultShippingAddress: 0,
-      }),
-      ...(this.model.currentFormValues['isDefaultBilling'] && {
-        defaultBillingAddress: 1,
-      }),
-    };
+      const data = {
+        email: this.model.currentFormValues.email,
+        password: this.model.currentFormValues.password,
+        dateOfBirth: this.model.currentFormValues.birthday,
+        firstName: this.model.currentFormValues.name,
+        lastName: this.model.currentFormValues.surname,
+        addresses: [
+          {
+            country: this.model.currentFormValues.country,
+            streetName: this.model.currentFormValues.street,
+            postalCode: this.model.currentFormValues['postalCode'],
+            city: this.model.currentFormValues.city,
+          },
+          {
+            country: this.model.currentFormValues['countryBilling'],
+            streetName: this.model.currentFormValues['streetBilling'],
+            postalCode: this.model.currentFormValues['postalCodeBilling'],
+            city: this.model.currentFormValues['cityBilling'],
+          },
+        ],
+        shippingAddresses: [0],
+        billingAddresses: this.model.currentFormValues['isShippingAsBilling'] ? [0] : [1],
 
-    try {
-      await authService.registerCustomer(data);
-      await new ModalGreeting('The account was created successfully').open();
-      route.navigate('/home');
-      LoginController.configureLogoutButton();
-      const auth = authService.getAuthenticatedStatus();
+        ...(this.model.currentFormValues['isDefaultShipping'] && {
+          defaultShippingAddress: 0,
+        }),
+        ...(this.model.currentFormValues['isDefaultBilling'] && {
+          defaultBillingAddress: 1,
+        }),
+      };
 
-      if (auth) {
-        localStorage.setItem('isLoggedPlants', 'true');
-      }
-    } catch (error) {
-      if (isCommercetoolsApiError(error)) {
-        handleApiError(error);
-      } else {
-        console.error('Unknown error', error);
+      try {
+        await authService.registerCustomer(data);
+        await new ModalGreeting('The account was created successfully').open();
+        route.navigate('/home');
+        LoginController.configureLogoutButton();
+        const auth = authService.getAuthenticatedStatus();
+
+        if (auth) {
+          localStorage.setItem('isLoggedPlants', 'true');
+        }
+      } catch (error) {
+        if (isCommercetoolsApiError(error)) {
+          handleApiError(error);
+        } else {
+          console.error('Unknown error', error);
+        }
+      } finally {
+        this.isActive = true;
       }
     }
   };
