+++ src/pages/layout/footer.scss
@@ -1,7 +1,7 @@
 .footer-logo__img {
   width: 32px;
   height: 32px;
-  background: url('../assets/images/rss-logo.svg') no-repeat center/cover;
+  background: url('/images/rss-logo.svg') no-repeat center/cover;
 }
 
 .footer-logo__link {
