+++ src/pages/not-found/not-found.ts
@@ -1,5 +1,5 @@
 import { CreateButton } from '../../components/button/create-button.ts';
-import type { IParameters } from '../../shared/models/interfaces/index.ts';
+import type { IParameters } from '../../shared/models/interfaces';
 import { CreateElement } from '../../shared/utils/create-element.ts';
 import { View } from '../view.ts';
