+++ src/pages/registration/billing-address-elements.ts
@@ -2,7 +2,7 @@ import Element from '../../components/element/element.ts';
 import { CreateInput } from '../../components/input/create-input.ts';
 import { Label } from '../../components/label/label.ts';
 import { COUNTRIES } from '../../shared/constants/messages-for-validator.ts';
-import type { IParameters } from '../../shared/models/interfaces/index.ts';
+import type { IParameters } from '../../shared/models/interfaces';
 import { CreateElement } from '../../shared/utils/create-element.ts';
 
 import './styles.scss';
