+++ src/pages/registration/credentials-elements.ts
@@ -1,7 +1,7 @@
 import Element from '../../components/element/element.ts';
 import { CreateInput } from '../../components/input/create-input.ts';
 import { Label } from '../../components/label/label.ts';
-import type { IParameters } from '../../shared/models/interfaces/index.ts';
+import type { IParameters } from '../../shared/models/interfaces';
 import { CreateElement } from '../../shared/utils/create-element.ts';
 
 import './styles.scss';
