+++ src/pages/registration/styles.scss
@@ -42,7 +42,7 @@
   width: 25px;
   height: 25px;
 
-  background-image: url('../../assets/images/close.svg');
+  background-image: url('../../../public/images/close.svg');
 }
 
 .group {
