+++ src/router/routes.constant.ts
@@ -1,6 +1,6 @@
 import type { IRoutes } from '../shared/models/interfaces';
 
-export const ROUTES: IRoutes = {
+export const ROUTES_URL: IRoutes = {
   MAIN: '/',
   MAIN_ALT: '/home',
   LOGIN: '/login',
