+++ src/shared/models/typeguards.ts/typeguards.ts
@@ -1,7 +1,7 @@
 import type { TokenStore } from '@commercetools/ts-client';
 
 import type { IFormValues, IResponse } from '../interfaces';
-import type { CommercetoolsApiError } from '../type/commercetools-api-errors';
+import type { CommercetoolsApiError } from '../type';
 
 export const isHTMLElement = (value: unknown): value is HTMLElement => value instanceof HTMLElement;
