+++ src/styles/scss/_fonts.scss
@@ -3,7 +3,7 @@
   font-weight: 400;
   font-style: normal;
   font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Regular.woff2') format('woff2');
+  src: url('/assets/fonts/CeraPro-Regular.woff2') format('woff2');
 }
 
 @font-face {
@@ -11,7 +11,7 @@
   font-weight: 500;
   font-style: normal;
   font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Medium.woff2') format('woff2');
+  src: url('/assets/fonts/CeraPro-Medium.woff2') format('woff2');
 }
 
 @font-face {
@@ -19,7 +19,7 @@
   font-weight: 700;
   font-style: normal;
   font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Bold.woff2') format('woff2');
+  src: url('/assets/fonts/CeraPro-Bold.woff2') format('woff2');
 }
 
 @font-face {
@@ -27,5 +27,5 @@
   font-weight: 900;
   font-style: normal;
   font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Black.woff2') format('woff2');
+  src: url('/assets/fonts/CeraPro-Black.woff2') format('woff2');
 }
