+++ src/styles/scss/_variables.scss
@@ -121,5 +121,5 @@ $container-lg: 960px;
 $container-xl: 1140px;
 $container-xxl: 1320px;
 
-$visibility-icon-on: url('../../assets/images/svg/visibility-on.svg');
-$visibility-icon-off: url('../../assets/images/svg/visibility-off.svg');
+$visibility-icon-on: url('../../../public/images/svg/visibility-on.svg');
+$visibility-icon-off: url('../../../public/images/svg/visibility-off.svg');
