--- src/styles/style.css
@@ -1,60 +1,0 @@
-@font-face {
-  font-family: 'Cera Pro';
-  font-weight: 400;
-  font-style: normal;
-  font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Regular.woff2') format('woff2');
-}
-
-@font-face {
-  font-family: 'Cera Pro';
-  font-weight: 500;
-  font-style: normal;
-  font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Medium.woff2') format('woff2');
-}
-
-@font-face {
-  font-family: 'Cera Pro';
-  font-weight: 700;
-  font-style: normal;
-  font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Bold.woff2') format('woff2');
-}
-
-@font-face {
-  font-family: 'Cera Pro';
-  font-weight: 900;
-  font-style: normal;
-  font-display: swap;
-  src: url('../../assets/fonts/CeraPro-Black.woff2') format('woff2');
-}
-
-* {
-  box-sizing: border-box;
-  margin: 0;
-  padding: 0;
-}
-
-:root {
-  display: flex;
-  place-items: center;
-  justify-content: center;
-  box-sizing: border-box;
-}
-
-html {
-  scroll-behavior: smooth;
-  font-size: 16px;
-}
-
-img {
-  max-width: 100%;
-  height: auto;
-}
-
-.not-found__message {
-  font-family: 'Cera Pro', sans-serif;
-  font-size: 20px;
-  font-weight: 900;
-}
